Remove NaNs from county shape data
The original coordinates for Jackson county had an NaN. I believe this
NaN is there to signify a break, where the following coordinates
before and after represnt two different boundaries. A closer look
show that to be the case, where the coordinates before representing
a small rectangular boundary. The NaNs were removed because it was
causing issues when used with Bokeh. Further, removing them did not
change the resulting county shape.